requirements.txt
# Web API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Core video processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Whisper for subtitles
openai-whisper>=20231117

# Translation (optional)
deep_translator>=1.11.0

# Original dependencies
requests
playwright
eventlet
schedule
cf_clearance
biliup
xhs
qrcode
loguru
apscheduler
f2
pywin32
ruamel.yaml
opencc-python-reimplemented
pysrt
pydub
watchdog
pyyaml
psutil
openai
dashscope